ai-agent/requirements.txt
fastapi
uvicorn
chromadb
google-genai
pydantic
python-dotenv
numpy
langgraph
langchain-core
